+++ node/Cargo.toml
@@ -41,6 +41,7 @@ pallet-configuration = { workspace = true, features = [ "std" ] }
 pallet-registrar-runtime-api = { workspace = true, features = [ "std" ] }
 services-payment-rpc = { workspace = true }
 stream-payment-rpc = { workspace = true }
+tc-service-container-chain = { workspace = true }
 tp-author-noting-inherent = { workspace = true, features = [ "std" ] }
 
 dc-orchestrator-chain-interface = { workspace = true }
@@ -112,8 +113,10 @@ cumulus-client-pov-recovery = { workspace = true }
 cumulus-client-service = { workspace = true }
 cumulus-primitives-core = { workspace = true }
 cumulus-relay-chain-interface = { workspace = true }
+
 [dev-dependencies]
 sp-panic-handler = { workspace = true }
+
 [build-dependencies]
 substrate-build-script-utils = { workspace = true }
 
@@ -134,6 +137,7 @@ runtime-benchmarks = [
 	"polkadot-service/runtime-benchmarks",
 	"sc-service/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
+	"tc-service-container-chain/runtime-benchmarks",
 ]
 try-runtime = [
 	"dancebox-runtime/try-runtime",
@@ -144,6 +148,7 @@ try-runtime = [
 	"polkadot-cli/try-runtime",
 	"polkadot-service/try-runtime",
 	"sp-runtime/try-runtime",
+	"tc-service-container-chain/try-runtime",
 ]
 
 fast-runtime = [ "dancebox-runtime/fast-runtime", "flashbox-runtime/fast-runtime" ]
